Swift: fix extraction of sources from `..`
