--- swift/extractor/infra/Path.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <filesystem>
-
-namespace codeql {
-std::filesystem::path getCodeQLPath(std::string_view path);
-}
