+++ swift/extractor/infra/file/Path.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <filesystem>
+
+namespace codeql {
+std::filesystem::path resolvePath(std::string_view path);
+}
