+++ swift/integration-tests/posix-only/frontend-invocations/Files.expected
@@ -5,4 +5,5 @@
 | E.swift:0:0:0:0 | E.swift |
 | F1.swift:0:0:0:0 | F1.swift |
 | F2.swift:0:0:0:0 | F2.swift |
+| G.swift:0:0:0:0 | G.swift |
 | file://:0:0:0:0 |  |
